Renamed string utils in MPE module
